+++ src/main/kotlin/app/revanced/patches/youtube/ad/general/resource/patch/GeneralAdsResourcePatch.kt
@@ -88,6 +88,20 @@ class GeneralAdsResourcePatch : ResourcePatch {
                 StringResource("revanced_adremover_compact_banner_enabled_summary_off", "Compact banners are shown")
             ),
             SwitchPreference(
+                "revanced_adremover_view_products",
+                StringResource("revanced_adremover_view_products_title", "Hide banner to view products"),
+                true,
+                StringResource("revanced_adremover_view_products_summary_on", "Banner is hidden"),
+                StringResource("revanced_adremover_view_products_summary_off", "Banner is shown")
+            ),
+            SwitchPreference(
+                "revanced_adremover_web_search_result",
+                StringResource("revanced_adremover_web_search_result_panel_title", "Hide web search results"),
+                true,
+                StringResource("revanced_adremover_web_search_result_summary_on", "Web search results are hidden"),
+                StringResource("revanced_adremover_web_search_result_summary_off", "Web search results are shown")
+            ),
+            SwitchPreference(
                 "revanced_adremover_movie",
                 StringResource("revanced_adremover_movie_enabled_title", "Hide movies section"),
                 true,